Move hooks out of lib/ into src/frontend/hooks/ for flatter structure
