+++ src/frontend/hooks/projects.ts
@@ -1,6 +1,6 @@
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import { useParams } from "react-router-dom";
-import { api } from "../api";
+import { api } from "../lib/api";
 import { unwrap } from "./util";
 
 export function useProjects() {
