+++ src/frontend/hooks/settings.ts
@@ -1,5 +1,5 @@
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
-import { api } from "../api";
+import { api } from "../lib/api";
 import { unwrap } from "./util";
 
 export function useProjectDoc(projectId: string | undefined) {
